// src/app/Models/menu-item.model.ts
export class MenuItem {
    MENU_ID: number;
    MENU_ID_LEVEL0: number;
    MENU_ID_LEVEL1: number;
    MENU_ID_LEVEL2: number;
    MENU_NAME: string;
    MENU_URL: string;
    ENABLED_FL: string;
    ENTRY_TS: Date;
    ENTRY_WORKER: string;
    ENTRY_USER: string;
    UPDATE_TS: Date;
    UPDATE_WORKER: string;
    UPDATE_USER: string;
    ROW_CHANGE_TS: Date;  
}